requirements.txt
asgiref==3.2.9
autopep8==1.6.0
Django==3.0.7
djangorestframework==3.13.1
Pillow==7.1.2
pycodestyle==2.8.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.2
